move track logic to room
