+++ src/components/TrackInfo.tsx
@@ -15,7 +15,6 @@ class TrackInfo extends React.Component<Props, State> {
     const { playerState, classes } = this.props;
     const track = playerState?.track_window?.current_track
     const img = track?.album?.images[0]?.url
-    console.log("IMG: ", img)
     const artist = track?.artists
       .map(artist => artist.name)
       .join(', ')
@@ -46,7 +45,7 @@ interface Props {
 interface State {}
 
 const mapStateToProps = (state: GlobalState) => ({
-  playerState: state.track.playerState,
+  playerState: state.room.playerState,
 });
 
 const mapDispatchToProps = { };
