+++ src/redux/logic.ts
@@ -1,9 +1,7 @@
 import userLogic from './user/logic'
-import trackLogic from './track/logic'
 import roomLogic from './room/logic'
 
 export default [
   ...userLogic,
-  ...trackLogic,
   ...roomLogic,
 ] as any
