+++ src/redux/room/logic.ts
@@ -14,12 +14,15 @@ import {
   auxPassedSuccess,
   syncPlayer,
   SYNC_PLAYER,
+  trackStatus,
+  updateTrack,
+  UPDATE_TRACK,
 } from "./actions";
-import { updateTrack, startListening } from "../track/actions";
+// import { updateTrack, startListening } from "../track/actions";
 import RoomClient from "../../room/client";
 import { PlayerState } from "../../spotify/types";
 import { Member } from "../../room/types";
-import spotify from "../../spotify";
+import spotify, { DEBOUNCE_RANGE } from "../../spotify";
 
 const createRoomLogic = createLogic({
   type: CREATE_ROOM,
@@ -39,9 +42,6 @@ const createRoomLogic = createLogic({
 const connectToRoomLogic = createLogic({
   type: CONNECT_TO_ROOM,
   async process({ getState, action }: ProcessOpts, dispatch, done) {
-    // start player
-    dispatch(startListening())
-
     const { roomname } = action.payload;
     const username = getState().user.name;
 
@@ -74,6 +74,18 @@ const syncPlayerLogic = createLogic({
   warnTimeout: 0,
   async process({ getState, action }: ProcessOpts, dispatch, done) {
     const { room, player } = action.payload;
+    player.addListener('player_state_changed', (playerState: PlayerState) => {
+      const room = getState().room.room
+      if(room === null){
+        // TODO: Better error handling here
+        throw new Error("Not connect to room")
+      }
+      if(room.leader === room.socket.id) {
+        room.updateTrack(playerState)
+      }
+      dispatch(trackStatus(playerState))
+    });
+
     room.onTrackUpdate = (data: PlayerState) => {
       const roomState = getState().room;
       if (roomState.userId !== roomState.leader) {
@@ -82,7 +94,7 @@ const syncPlayerLogic = createLogic({
     };
     room.onMemberAdded = async (member: Member) => {
       dispatch(memberAdded(member))
-      const playerState = await getState().track.player?.getCurrentState()
+      const playerState = await getState().room.player?.getCurrentState()
       if(playerState) {
         room.updateTrack(playerState)
       }
@@ -91,6 +103,43 @@ const syncPlayerLogic = createLogic({
     room.onAuxPassed = (id: string) => dispatch(auxPassed(id));
   },
 });
+
+const updateTrackLogic = createLogic({
+  type: UPDATE_TRACK,
+  async process({ getState, action }: ProcessOpts, dispatch, done) {
+    const info: PlayerState = action.payload
+
+    const token = getState().user.token
+    if(!token){
+      throw new Error("User not logged in")
+    }
+
+
+    const { player } = getState().room
+    if(!player) {
+      throw new Error("No player")
+    }
+    const currState = await player.getCurrentState()
+
+    const infoUri = info.track_window.current_track.uri
+    const stateUri = currState?.track_window?.current_track?.uri
+
+    const progressDiff = Math.abs(info.position - (currState?.position || 0))
+    if(info.paused && !currState?.paused){
+      await player.pause()
+    } else if(!info.paused && currState?.paused) {
+      await player.resume()
+    }
+
+    if(infoUri !== stateUri || progressDiff > DEBOUNCE_RANGE) {
+      await spotify.changeTrack(token, infoUri, info.position)
+    }
+    dispatch(trackStatus(info))
+    done();
+  }
+});
+
+
 
 const passAuxLogic = createLogic({
   type: PASS_AUX,
@@ -104,26 +153,26 @@ const passAuxLogic = createLogic({
   },
 });
 
-const auxPassedLogic = createLogic({
-  type: AUX_PASSED,
-  async process({ getState, action }: ProcessOpts, dispatch, done) {
-    // const { userId, leader } = getState().room;
-    const newLeader = action.payload.id;
-    // if (userId === leader) {
-      // dispatch(stopListening());
-    // }
-    // if (userId === newLeader) {
-    //   dispatch(startListening());
-    // }
-    dispatch(auxPassedSuccess(newLeader));
-    done();
-  },
-});
+// const auxPassedLogic = createLogic({
+//   type: AUX_PASSED,
+//   async process({ getState, action }: ProcessOpts, dispatch, done) {
+//     // const { userId, leader } = getState().room;
+//     const newLeader = action.payload.id;
+//     // if (userId === leader) {
+//       // dispatch(stopListening());
+//     // }
+//     // if (userId === newLeader) {
+//     //   dispatch(startListening());
+//     // }
+//     dispatch(auxPassedSuccess(newLeader));
+//     done();
+//   },
+// });
 
 export default [
   createRoomLogic,
   connectToRoomLogic,
   syncPlayerLogic,
+  updateTrackLogic,
   passAuxLogic,
-  auxPassedLogic,
 ];
